+++ poem-openapi/src/param/query.rs
@@ -55,17 +55,21 @@ impl<'a, T: ParseFromParameter> ApiExtractor<'a> for Query<T> {
         _body: &mut RequestBody,
         param_opts: ExtractParamOptions<Self::ParamType>,
     ) -> Result<Self, ParseRequestError> {
-        let value = request
+        let mut values = request
             .extensions()
             .get::<UrlQuery>()
-            .and_then(|query| query.0.get(param_opts.name).map(|s| s.as_str()));
-        let value = match (value, &param_opts.default_value) {
-            (Some(value), _) => Some(value),
-            (None, Some(default_value)) => return Ok(Self(default_value())),
-            (None, _) => None,
-        };
+            .unwrap()
+            .get_all(param_opts.name)
+            .peekable();
 
-        ParseFromParameter::parse_from_parameter(value)
+        match &param_opts.default_value {
+            Some(default_value) if values.peek().is_none() => {
+                return Ok(Self(default_value()));
+            }
+            _ => {}
+        }
+
+        ParseFromParameter::parse_from_parameters(values)
             .map(Self)
             .map_err(|err| ParseRequestError::ParseParam {
                 name: param_opts.name,